% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_cell_viewer.R
\name{training_cell_viewer}
\alias{training_cell_viewer}
\title{Interactively Visualize and Subset UMAPs}
\usage{
training_cell_viewer(umap)
}
\arguments{
\item{umap}{a \code{tibble()} with}
}
\description{
Interactively Visualize and Subset UMAPs
}
