% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_hallmarks.R
\name{training_hallmarks}
\alias{training_hallmarks}
\title{Helper Functions to Process 'Hallmarks of Cancer' MSigDB Gene Sets}
\usage{
training_hallmarks()
}
\value{
\code{training_hallmarks()} creates a tibble with columns 'gene'
(Ensembl identifiers), 'set', and 'description' (a link to a
description of the set on the MSigDB web site).
}
\description{
\code{training_hallmarks()} retrieves the 'Hallmarks of
Cancer' geen sets from MSigDB.
}
\details{
As is often the case, the Hallmarks of Cancer gene sets
use gene identifiers from the NCBI ('Entrez' gene identifiers),
but our data uses gene identifiers from Ensembl; this function
accomplishes the tedious task of translating between gene sets
using the Bioconductor 'org.Hs.eg.db' data resource and
annotation functions in the AnnotationDbi pacakge. The mapping
between identifiers is not 1:1, so the number of genes in each
Hallmark set differs between Ensembl and Entrez identifiers.
}
\examples{
training_hallmarks()

}
