% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial_cell_viewer.R
\name{tutorial_cell_viewer}
\alias{tutorial_cell_viewer}
\title{Interactively Visualize and Subset UMAPs}
\usage{
tutorial_cell_viewer(umap)
}
\arguments{
\item{umap}{a \code{tibble()} with}
}
\description{
Interactively Visualize and Subset UMAPs
}
