% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxg_dataset.R
\name{cxg_dataset}
\alias{cxg_dataset}
\alias{print.cxg_dataset}
\title{Summarize CxG Information About a Dataset}
\usage{
cxg_dataset(dataset_id, cellxgene_db = db())

\method{print}{cxg_dataset}(x, ...)
}
\arguments{
\item{dataset_id}{character(1) dataset identifier, as returned by,
e.g., \code{datasets()}.}
}
\description{
Summarize CxG Information About a Dataset
}
\examples{
dataset <- "de985818-285f-4f59-9dbd-d74968fddba3"
ds <- cxg_dataset(dataset)
ds

}
