% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_cxg_dataset.R
\name{training_cxg_dataset}
\alias{training_cxg_dataset}
\alias{print.cxg_dataset}
\title{Summarize CxG Information About a Dataset}
\usage{
training_cxg_dataset(dataset_id)

\method{print}{cxg_dataset}(x, ...)
}
\arguments{
\item{dataset_id}{character(1) dataset identifier, as returned by,
e.g., \code{datasets()}.}

\item{x}{A \code{cxg_dataset} object resulting from a call to
\code{training_cxg_dataset()}.}

\item{...}{Additional arguments (to \code{print.cxg_dataset()}); ignored.}
}
\description{
Summarize CxG Information About a Dataset
}
\examples{
dataset <- "de985818-285f-4f59-9dbd-d74968fddba3"
ds <- training_cxg_dataset(dataset)
ds

}
