% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_read_h5ad.R
\name{training_read_h5ad_as_sce}
\alias{training_read_h5ad_as_sce}
\alias{training_read_h5ad_as_seurat}
\title{Read 'h5ad' Files as R / Bioconductor Objects}
\usage{
training_read_h5ad_as_sce(h5ad_file_path)

training_read_h5ad_as_seurat(h5ad_file_path)
}
\arguments{
\item{h5ad_file_path}{\code{character(1)} path to the \code{.h5ad} file.}
}
\value{
\code{training_read_h5ad_as_sce()} returns an R / Bioconductor
SingleCellExperiment object. The assay data \code{X} is the (sparse)
matrix named 'X' in the h5ad file.
}
\description{
\code{read_h5ad_as_sce()} uses the Python h5ad module and R
'anndata' package to import the \code{.h5ad} objects into R /
Bioconductor 'SingleCellExperiment'.
}
